requirements.txt
# requirements.txt

# Example packages
flask
torch
transformers
accelerate
python-dotenv
bitsandbytes
